fix: refactor folder structure to be more intuitive
