+++ src/styledJsx/babel.ts
@@ -1,0 +1,8 @@
+import { TransformOptions } from '@babel/core'
+
+export const configureStyledJsxTransforms = (
+  config: TransformOptions
+): TransformOptions => ({
+  ...config,
+  plugins: [...(config.plugins ?? []), 'styled-jsx/babel']
+})
