+++ src/styledJsx/webpack.ts
@@ -1,0 +1,5 @@
+import { Configuration as WebpackConfig } from 'webpack'
+import { addScopedAlias } from '../utils'
+
+export const configureStyledJsx = (baseConfig: WebpackConfig): void =>
+  addScopedAlias(baseConfig, 'styled-jsx')
